requirements.txt
python-telegram-bot==13.15
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.0.1
apscheduler==3.11.0
pytz==2025.2
